+++ cats/plugins.py
@@ -1,5 +1,7 @@
 from typing import Type, Union
 
+import orjson
+
 try:
     from django.db.models import QuerySet, Model
 except ImportError:
@@ -30,6 +32,7 @@ __all__ = [
     'Form',
     'scheme_load',
     'scheme_dump',
+    'scheme_json',
 ]
 
 Scheme = Union[Type[BaseModel], Type[BaseSerializer], Type[ModelSchema]]
@@ -50,6 +53,10 @@ class DRF:
             assert isinstance(data, s)
             return data.data
         return s(data, many=many).data
+
+    @classmethod
+    def json(cls, s: Type[BaseSerializer], data, *, many: bool = False, plain: bool = False) -> bytes:
+        return orjson.dumps(cls.dump(s, data, many=many, plain=plain))
 
 
 class Pydantic:
@@ -74,6 +81,22 @@ class Pydantic:
             data = s.parse_obj(data)
         return data.dict()
 
+    @classmethod
+    def json(cls, s: Type[BaseModel], data, *, many: bool = False, plain: bool = False) -> bytes:
+        if many:
+            return b'[' + b','.join(cls.json(s, i, many=False, plain=plain) for i in data) + b']'
+
+        if plain:
+            assert isinstance(data, s)
+        elif isinstance(data, (QuerySet, Model)):
+            data = s.from_orm(data)
+        else:
+            data = s.parse_obj(data)
+        res = data.json()
+        if isinstance(res, bytes):
+            return res
+        return res.encode('utf-8')
+
 
 class Djantic(Pydantic):
     @classmethod
@@ -88,7 +111,24 @@ class Djantic(Pydantic):
             assert isinstance(data, s)
         else:
             data = s.from_django(data)
-        return data.dict()
+        return data.json()
+
+    @classmethod
+    def json(cls, s: Type[ModelSchema], data, *, many: bool = False, plain: bool = False) -> bytes:
+        if many:
+            return b'[' + b','.join(cls.json(s, i, many=False, plain=plain) for i in data) + b']'
+
+        if not isinstance(data, (QuerySet, Model)):
+            return super().json(s, data, many=many, plain=plain)
+
+        if plain:
+            assert isinstance(data, s)
+        else:
+            data = s.from_django(data)
+        res = data.json()
+        if isinstance(res, bytes):
+            return res
+        return res.encode('utf-8')
 
 
 def _resolve_scheme_type(scheme: Scheme):
@@ -108,3 +148,7 @@ def scheme_load(scheme: Scheme, data, *, many: bool = False, plain: bool = False
 
 def scheme_dump(scheme: Scheme, data, *, many: bool = False, plain: bool = False):
     return _resolve_scheme_type(scheme).dump(scheme, data, many=many, plain=plain)
+
+
+def scheme_json(scheme: Scheme, data, *, many: bool = False, plain: bool = False) -> bytes:
+    return _resolve_scheme_type(scheme).json(scheme, data, many=many, plain=plain)